Functional state on short interval.
Sufficient for experimentation. Not ready for regular deployment
(disabling on startup). Interval is short and does not add guilds
to tracked guild set in Redis. Need to find a way to account for the
length and diversity of messages containing emojis. Regex filter?
Seems inefficient but I might not have options.